Replace Meeseeks notifications with platform-specific scheduling
