+++ composeApp/src/commonMain/kotlin/com/hansholz/bestenotenapp/notifications/GradeNotificationEngine.kt
@@ -1,0 +1,145 @@
+package com.hansholz.bestenotenapp.notifications
+
+import androidx.compose.runtime.mutableStateOf
+import com.hansholz.bestenotenapp.api.BesteSchuleApi
+import com.hansholz.bestenotenapp.api.createHttpClient
+import com.hansholz.bestenotenapp.api.models.Grade
+import com.hansholz.bestenotenapp.api.models.GradeCollection
+import com.hansholz.bestenotenapp.security.AuthTokenManager
+import com.mmk.kmpnotifier.notification.NotifierManager
+import com.russhwolf.settings.Settings
+import com.russhwolf.settings.getStringOrNull
+import io.ktor.client.plugins.ClientRequestException
+import kotlinx.coroutines.CancellationException
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.decodeFromString
+import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.Json
+import kotlin.time.Duration.Companion.minutes
+
+internal enum class GradeNotificationOutcome {
+    Success,
+    Retry
+}
+
+internal object GradeNotificationEngine {
+    private const val KEY_KNOWN_GRADE_IDS = "gradeNotificationsKnownGradeIds"
+    private const val MIN_INTERVAL_MINUTES = 15L
+
+    private val json = Json { ignoreUnknownKeys = true }
+    private val settings = Settings()
+    private val authTokenManager = AuthTokenManager()
+
+    fun minimumIntervalMinutes(): Long = MIN_INTERVAL_MINUTES
+
+    fun isEnabled(): Boolean = settings.getBoolean(GradeNotifications.KEY_ENABLED, false)
+
+    fun getIntervalMinutes(): Long =
+        settings.getLong(GradeNotifications.KEY_INTERVAL_MINUTES, GradeNotifications.DEFAULT_INTERVAL_MINUTES)
+
+    fun shouldSchedule(): Boolean = isEnabled() && hasCredentials()
+
+    fun clearKnownGrades() {
+        settings.remove(KEY_KNOWN_GRADE_IDS)
+    }
+
+    suspend fun runCheck(): GradeNotificationOutcome {
+        if (!shouldSchedule()) return GradeNotificationOutcome.Success
+
+        val studentId = settings.getStringOrNull("studentId") ?: return GradeNotificationOutcome.Success
+        val token = authTokenManager.getToken()
+        if (token.isNullOrBlank()) return GradeNotificationOutcome.Success
+
+        val httpClient = createHttpClient()
+        return try {
+            val authState = mutableStateOf(token)
+            val studentState = mutableStateOf(studentId)
+            val api = BesteSchuleApi(httpClient, authState, studentState)
+            val includes = listOf("grades", "interval", "subject", "grades.histories")
+            val collections = fetchAllCollections(api, includes)
+            val currentIds = collections.flatMap { it.grades.orEmpty() }.map { it.id }.toSet()
+
+            val knownIds = loadKnownGradeIds()
+            val newIds = currentIds - knownIds
+
+            if (newIds.isNotEmpty()) {
+                val newGrades = collections.flatMap { collection ->
+                    collection.grades.orEmpty()
+                        .filter { it.id in newIds }
+                        .map { it to collection }
+                }.sortedBy { it.second.givenAt }
+                notifyNewGrades(newGrades)
+            }
+
+            storeKnownGradeIds(currentIds)
+            GradeNotificationOutcome.Success
+        } catch (e: CancellationException) {
+            throw e
+        } catch (e: ClientRequestException) {
+            if (e.response.status.value == 401) {
+                GradeNotificationOutcome.Success
+            } else {
+                GradeNotificationOutcome.Retry
+            }
+        } catch (e: Exception) {
+            GradeNotificationOutcome.Retry
+        } finally {
+            httpClient.close()
+        }
+    }
+
+    private fun hasCredentials(): Boolean {
+        val studentId = settings.getStringOrNull("studentId")
+        val token = authTokenManager.getToken()
+        return !studentId.isNullOrBlank() && !token.isNullOrBlank()
+    }
+
+    private suspend fun fetchAllCollections(
+        api: BesteSchuleApi,
+        includes: List<String>
+    ): List<GradeCollection> {
+        val result = mutableListOf<GradeCollection>()
+        val firstPage = api.collectionsIndex(include = includes)
+        result += firstPage.data
+        val lastPage = firstPage.meta?.lastPage ?: 1
+        if (lastPage > 1) {
+            for (page in 2..lastPage) {
+                result += api.collectionsIndex(include = includes, page = page).data
+            }
+        }
+        return result
+    }
+
+    private fun notifyNewGrades(entries: List<Pair<Grade, GradeCollection>>) {
+        val notifier = NotifierManager.getLocalNotifier()
+        entries.forEach { (grade, collection) ->
+            val title = collection.name?.takeIf { it.isNotBlank() } ?: "Neue Note"
+            val subject = collection.subject?.name ?: grade.subject?.name ?: "Unbekanntes Fach"
+            val body = buildString {
+                append(subject)
+                if (grade.value.isNotBlank()) {
+                    append(" – ")
+                    append(grade.value)
+                }
+            }
+            notifier.notify {
+                id = grade.id
+                this.title = title
+                this.body = body
+            }
+        }
+    }
+
+    private fun loadKnownGradeIds(): Set<Int> {
+        val raw = settings.getStringOrNull(KEY_KNOWN_GRADE_IDS) ?: return emptySet()
+        return runCatching { json.decodeFromString<KnownGrades>(raw).ids.toSet() }.getOrElse { emptySet() }
+    }
+
+    private fun storeKnownGradeIds(ids: Set<Int>) {
+        val payload = json.encodeToString(KnownGrades(ids.toList()))
+        settings.putString(KEY_KNOWN_GRADE_IDS, payload)
+    }
+
+    @Serializable
+    private data class KnownGrades(val ids: List<Int>)
+}
