+++ iosApp/iosApp/Info.plist
@@ -46,16 +46,6 @@
         <string>UIInterfaceOrientationLandscapeLeft</string>
         <string>UIInterfaceOrientationLandscapeRight</string>
     </array>
-    <key>BGTaskSchedulerPermittedIdentifiers</key>
-    <array>
-        <string>dev.mattramotar.meeseeks.task.refresh</string>
-        <string>dev.mattramotar.meeseeks.task.processing</string>
-    </array>
-    <key>UIBackgroundModes</key>
-    <array>
-        <string>fetch</string>
-        <string>processing</string>
-    </array>
     <key>CFBundleLocalizations</key>
     <array>
         <string>de</string>
